requirements.txt

#devops
black==22.3.0
pytest==7.4.0
pytest-cov==4.0.0
nbval==0.11.0
pylint==2.15.3
#rust based linter
#ruff==0.0.284

#analysis & data viz
pandas
matplotlib
numpy
requests
databricks-sql-connector
python-dotenv